Runtime error function in runtime.c
Implicit runtime error is necessary as a debugging tool, for the
generated LLVM code. Previously the prim_int_print used to report
errors, but as the non-used externals are removed from the generated
code, this bug was releived and a quick fix is to have an explicit
runime_error function in the runtime.c module.

This should be optional at the final form of the llvm codegen.